backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==1.10.13
python-dotenv==1.0.1
mangum>=0.17.0
email-validator==2.1.0
